+++ src/scattermind/system/graph/args.py
@@ -23,10 +23,11 @@ from scattermind.system.names import (
 )
 from scattermind.system.readonly.access import DataAccess
 from scattermind.system.torch_util import DTypeName, get_dtype_name
-from scattermind.system.util import as_int_list, as_shape
+from scattermind.system.util import as_int_list, as_shape, to_bool
 
 
 ArgType = Literal[
+    "bool",
     "str",
     "int",
     "float",
@@ -43,7 +44,8 @@ ArgType = Literal[
 
 
 ArgumentType = (
-    str
+    bool
+    | str
     | int
     | float
     | list[int]
@@ -52,10 +54,12 @@ ArgumentType = (
     | DataInfoJSON
     | DTypeName
     | DataAccess
+    | None
 )
 """Available argument types as python types as seen in the JSON."""
 ArgumentRetType = (
-    str
+    bool
+    | str
     | int
     | float
     | list[int]
@@ -110,50 +114,93 @@ class NodeArg:
         return self._arg
 
     @overload
-    def get(self, type_name: Literal["str"]) -> str:
+    def get(
+            self,
+            type_name: Literal["bool"],
+            default: bool | None = None) -> bool:
         ...
 
     @overload
-    def get(self, type_name: Literal["int"]) -> int:
+    def get(
+            self,
+            type_name: Literal["str"],
+            default: str | None = None) -> str:
         ...
 
     @overload
-    def get(self, type_name: Literal["float"]) -> float:
+    def get(
+            self,
+            type_name: Literal["int"],
+            default: int | None = None) -> int:
         ...
 
     @overload
-    def get(self, type_name: Literal["list_int"]) -> list[int]:
+    def get(
+            self,
+            type_name: Literal["float"],
+            default: float | None = None) -> float:
         ...
 
     @overload
-    def get(self, type_name: Literal["shape"]) -> list[int | None]:
+    def get(
+            self,
+            type_name: Literal["list_int"],
+            default: list[int] | None = None) -> list[int]:
         ...
 
     @overload
-    def get(self, type_name: Literal["ushape"]) -> list[int]:
+    def get(
+            self,
+            type_name: Literal["shape"],
+            default: list[int | None] | None = None) -> list[int | None]:
         ...
 
     @overload
-    def get(self, type_name: Literal["format"]) -> DataFormatJSON:
+    def get(
+            self,
+            type_name: Literal["ushape"],
+            default: list[int] | None = None) -> list[int]:
         ...
 
     @overload
-    def get(self, type_name: Literal["info"]) -> DataInfo:
+    def get(
+            self,
+            type_name: Literal["format"],
+            default: DataFormatJSON | None = None) -> DataFormatJSON:
         ...
 
     @overload
-    def get(self, type_name: Literal["graph"]) -> QualifiedGraphName:
+    def get(
+            self,
+            type_name: Literal["info"],
+            default: DataInfo | None = None) -> DataInfo:
         ...
 
     @overload
-    def get(self, type_name: Literal["dtype"]) -> DTypeName:
+    def get(
+            self,
+            type_name: Literal["graph"],
+            default: QualifiedGraphName | None = None) -> QualifiedGraphName:
         ...
 
     @overload
-    def get(self, type_name: Literal["data"]) -> DataAccess:
+    def get(
+            self,
+            type_name: Literal["dtype"],
+            default: DTypeName | None = None) -> DTypeName:
         ...
 
-    def get(self, type_name: ArgType) -> ArgumentRetType:
+    @overload
+    def get(
+            self,
+            type_name: Literal["data"],
+            default: DataAccess | None = None) -> DataAccess:
+        ...
+
+    def get(
+            self,
+            type_name: ArgType,
+            default: ArgumentRetType | None = None) -> ArgumentRetType:
         """
         Parses the argument and converts it into the internal type.
 
@@ -161,14 +208,23 @@ class NodeArg:
             type_name (ArgType): The name of the argument type. This should be
                 a literal string to make type checking infer the return
                 type correctly.
+            default (ArgumentRetType | None): The default value. If None, no
+                default value is provided.
 
         Raises:
             ValueError: If the argument could not be parsed.
+            KeyError: If the argument was not set and no default is provided.
 
         Returns:
             ArgumentRetType: The argument converted to the internal type.
         """
         val = self._arg
+        if val is None:
+            if default is None:
+                raise KeyError(f"argument {self._name} not set")
+            return default
+        if type_name == "bool":
+            return to_bool(cast(str, val))
         if type_name == "str":
             return f"{val}"
         if type_name == "graph":